JohnBryce-Homework/2025.12.14/requirements.txt
asgiref==3.11.0
Django==4.2.27
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
Pillow==10.1.0
PyJWT==2.10.1
pytz==2025.2
rest-framework-simplejwt==0.0.2
sqlparse==0.5.4
typing_extensions==4.15.0
